fix(sidekick/swift): `oneof` query parameters
The generated code to access `oneof` fields as query parameters was
wrong. Each `oneof` is generated as an enum, and a single field
containing all (though only one at a time is set) the fields in the
`oneof` group.

The generated code should access field through the `enum` field.
