+++ internal/sidekick/swift/generate_service_swift_test.go
@@ -172,91 +172,6 @@ func TestGenerateService_Delegation(t *testing.T) {
 	}
 }
 
-func TestGenerateService_StubStructure(t *testing.T) {
-	outDir := t.TempDir()
-
-	request := &api.Message{
-		Name:    "Request",
-		ID:      ".test.Request",
-		Package: "test",
-	}
-	response := &api.Message{
-		Name:    "Response",
-		ID:      ".test.Response",
-		Package: "test",
-	}
-	service := &api.Service{
-		Name:    "Protocol",
-		ID:      ".test.Prototocol",
-		Package: "test",
-		Methods: []*api.Method{
-			{
-				Name:         "GetThing",
-				ID:           ".test.IAM.CreateRole",
-				InputTypeID:  ".test.Request",
-				InputType:    request,
-				OutputTypeID: ".test.Response",
-				OutputType:   response,
-				PathInfo: &api.PathInfo{
-					Bindings: []*api.PathBinding{{Verb: "GET", PathTemplate: &api.PathTemplate{}}},
-				},
-			},
-		},
-	}
-
-	model := api.NewTestAPI([]*api.Message{request, response}, nil, []*api.Service{service})
-	model.PackageName = "google.cloud.test.v1"
-
-	cfg := &parser.ModelConfig{
-		Codec: map[string]string{
-			"copyright-year": "2038",
-		},
-	}
-
-	swiftCfg := swiftConfig(t, []config.SwiftDependency{
-		{
-			Name:       "SomeTestPackage",
-			ApiPackage: "test",
-		},
-	})
-	if err := Generate(t.Context(), model, outDir, cfg, swiftCfg); err != nil {
-		t.Fatal(err)
-	}
-
-	filename := filepath.Join(outDir, "Sources", "GoogleCloudTestV1", "Clients", "ProtocolStub.swift")
-	content, err := os.ReadFile(filename)
-	if err != nil {
-		t.Fatal(err)
-	}
-	contentStr := string(content)
-
-	got := extractBlock(t, contentStr, `  protocol ProtocolStub {`, "\n"+`  }`)
-	want := `  protocol ProtocolStub {
-    func getThing(
-    request: Request, options: GoogleCloudGax.RequestOptions
-) async throws -> SomeTestPackage.Response
-
-  }`
-	if diff := cmp.Diff(want, got); diff != "" {
-		t.Errorf("mismatch (-want +got):\n%s", diff)
-	}
-
-	got = extractBlock(t, contentStr, `  class ProtocolTransport: `, `HTTPClient`)
-	want = `  class ProtocolTransport: ProtocolStub {
-    let inner: GoogleCloudGax.HTTPClient`
-	if diff := cmp.Diff(want, got); diff != "" {
-		t.Errorf("mismatch (-want +got):\n%s", diff)
-	}
-
-	got = extractBlock(t, contentStr, `return try GoogleCloudWkt._ProtoJSONDecoder()`, ", from: data)\n    }")
-	want = `return try GoogleCloudWkt._ProtoJSONDecoder().decode(
-        SomeTestPackage.Response.self, from: data)
-    }`
-	if diff := cmp.Diff(want, got); diff != "" {
-		t.Errorf("mismatch (-want +got):\n%s", diff)
-	}
-}
-
 func TestGenerateService_SnippetFiles(t *testing.T) {
 	outDir := t.TempDir()
